Snapshot every commit with full detail, stored as weekly bases and patches
Every first-parent commit since package adoption has a detail file: weekly
commits and milestones hold the full detail, the rest a patch of the units
and stylesheets that changed since the base. Units store only their own facts
(hits, classes, imports, route); closures, blockers and route hits are
derived on load. Both commit pickers are searchable comboboxes over every
snapshot; the comparison defaults to a week earlier. Paths are relative to
src/main/webapp. The brief, llms.txt and history view are removed.

+++ src/features/migrations/lockable.tsx
@@ -9,7 +9,6 @@ import {
   TableRow,
 } from '@/components/ui/table'
 import { lockEntries, sourceUrl, type Detail } from './model'
-import { short } from './format'
 import { copyText } from './clipboard'
 
 function copyEntries(dirs: string[]) {
@@ -59,7 +58,7 @@ export function LockableTable({
                   className="underline underline-offset-4 break-all"
                   href={sourceUrl(commit, dir)}
                 >
-                  {short(dir)}
+                  {dir}
                 </a>
               </TableCell>
               <TableCell className="text-right tabular-nums">{units}</TableCell>
@@ -67,7 +66,7 @@ export function LockableTable({
                 <Button
                   variant="ghost"
                   size="icon"
-                  aria-label={`Copy lock entries for ${short(dir)}`}
+                  aria-label={`Copy lock entries for ${dir}`}
                   onClick={() => copyEntries([dir])}
                 >
                   <Copy aria-hidden="true" />
